improve UX by triggering tooltip on preference card title hover
Preference cards include a small (i) info icon that triggers some modal tooltip text
to appear when hovering (or tapping on mobile) that describes in detail the origins and nature of the preference under consideration.

With this fix, we extend the tooltip trigger area to include the intire mdl card title text.
New users may this the text area is a button an be prone to click it.  So make something useful appear that may be helpful.