improve UX by giving users guidance on using sliders in tooltip preference text
The app could use more context for new users, especially when they advance to the preference card page.

We may even need to offer a 'guided-option' for first-timers.  This could employ UI elements like a stepper.  For now, though, at least provide info in the tooltip text on using the slider elements to specify their preferences.