improve UX by tweaking text in header of preference page
Previously, we had 'Select preferences' in the header of the preference page.

But this may imply to the user that first they need to select preferences before dialing in the actualy levels of a given attribute via the sliders.

Since sliders are heavily featured in the UI, it's probably better to encourage the user to /adjust/ the current slider positions rather than select them.

The word 'adjust' is a bit more evocative of the action the user needs to take.