add model primitives for getting, setting and validating app state using localStorage api pull maxResults from localStorage if available and validated
I'm on a mission to persist application state to localStorage when appropriate.

With this commit, I define a structure called localState that will grow
to encompass all the state we want to persist on the client-side.

For now, it only includes the 'maxResults' value which tells the view
how many cities to feature on the results page.

I also rollout primitives for mutating localStorage in a safe way.