ux: refine language around preferences and results
Use 'Your priorities' instead of 'Your values'.

The term 'values' is a bit more expansive and intrinsic to how a person
self-identifies whereas 'priorities' is a bit more pragmatic in nature.
This app is about pragmatism.

Use 'Your best bets ...' instead of 'Consider these ...'

The app is providing something of value in the way of a ranked list of cities.
However, the term 'consider' is a bit detached and circumspect whereas a phrase
like 'Your best bets ..." aligns with verbiage in the splash screen and reflects
a more confident, value-add tone.