fix political stats for Detroit, MI
The prevailing politics data associated with Detroit, MI in the
static model was from Monroe County, MI but should be from Wayne County.

This is a bug that rippled in from City Rank (city-metrics.js).
The remainder of the model should be audited for consistency between
city and engulfing county.

I happened to notice this while running some usability testing and noticed
Detroit was asserted to be predominanty republican which I simply
know is not the case from all the PBS and NPR I consume.