design settings submenu; add (disabled) help menu item placeholder
Continuing with theme of fleshing out the menu and submenues.

For the settings, I'm adding an edit option rather than trying to
allow for inline editing of settings from within the submenu.

The plan is to add a dedicated settings page where I can easily
add a form and input elements to mutate app settings.