fix disappearing menu-drawer content seen with landscape orientations on mobile
I notice with an expanded 'Settings ...' submenu in a landscape
orientation on mobile devices, the menu items blip on then off the screen
annoyingly.

It appears the .mdl-menu__container class has a z-index of -1
that kicks in, rendering the submenu items invisible.

Unset the z-index in this scenario to restore visibility of settings items.
This also allows the menu to scroll nicely for content that is clipped
by the viewport.