fix edit settings top-cities drop-down selection list
Stuff that is working in the emulator in the dev environment
fails on a real device.

Try being more specific about the selector.  Content of drop-down
list is blipping behind the menu-outline.

These are the kinds of bugs that would probably get some love
in Material Components, but are neglected in MDL since it is
out of service.