refactor the settings code for i18n readiness
The roll-my-own dropdown-select elements used in the settings page
was based upon left-aligned geometries with hardcoded margins tied
to language-specific label strings.

With this commit, we make my input element more resilient to the
variable length of the strings to the left of the dropdown select
element.  (There's some implicit bias here for ltr languages.)

Also, take the opportunity to DRY some of the code and move some
related inline styles to classes in the our app's css file.

Finally, I'm removing the <form> tags wrappering the settings elements.
I didn't really have a form handler and clicking one of the controls
was triggering a redirect to the landing page.  Settings are pretty
light-weight at the moment and I'm okay with not having a Submit
button especially since we only have three settings and two of those
are placeholders for more aspirational features that are effectively
disabled for now (internationalization and country selection).