ux: settings dropdown menu should open /under/ active setting
Improve the ux by being more conventional with alignment behavior of
dropdown selection items.  These should be presented under the
active selection, not below and shifted right.

Also, use a shade of the primary background color in the visible
dropdown menu to better indicate when the dropdown is present.
Prior, the background was white and blending into the surrounding
page.