Allow the settings page to mutate number of top cities to report
With the settings view looking more spiffy, it's now time to add
a call to a setter that mutates state based upon the selected value
in the dropdown menus.

Here we update the number of top cities reported in the results page.

Still need to add 'reset to defaults' and 'clear cache' buttons on the settings page.
Also, we're not yet /restoring/ state from localStorage on startup.