add StaticModel.getScalarRange entrypoint
Doing battle against some hardcoded properties in the static model (like min/max affordability).

Adding a function that iterates over the array of city data and returns
the min and max values of a given numeric scalar property (like affordability or happiness).

This will allow us to upgrade the dataset more easily without having to manually
recompute the range and midpoint value of a given property, something the view needs to
configure preference sliders, for example.