refactor local persistence service to its own object move maxResults from city model to settings (view) model
The code that was persisting to localStorage was scattered in with the
city model.  For now, I'm going to let the controller decide when to persist
state.  In this commit, I implement a nicely-scoped persistence
api.

I'm also on a quest to introduce view models and make the M in our MVC more
canonical.  I'm starting by making the model a bit more granular
by separating Model into:

  CityModel and
  SettingsModel

Eventually I'll add a PrioritiesModel (for that view) and make all the
relevant models implement an Observable pattern so changes of
state can be pushed as notifications (through render callbacks) to the views
and any other interested layer.