integrate priorities model with view and controller
I'm moving state management for user priorities from the
view to the priorities model. This is more canonical from an MVC perspective.

Specifically, I'm killing off view.userPrefs.  The view now becomes a "getter" of the
state ensconsed within in the priorities object.

The controller becomes a "setter" of state in the priorities object
in response to slider events (or form input activity, more generally).

Strictly speaking, I'm composing the priorities model within the
controller which then exposes related interface callbacks to the view
via the view's constructor.

This will facilitate scaling up to a full stack implementations where we
introduce notion of session management along with persistence of state
to some backend database.