code complete on landing page view model
Removing some hardcoded strings from the view and pushing them to the model.

The ModelLanding class is essentially a string table of locale-sensitive
presentation text required by the splash screen.

It provides getter access to things like:

  application name
  slogan
  call-to-action blurb

I've coded it in a way that lends itself to simple localization if we want
to support that later.