have one source of truth for locale (in ModelSettings) fix tooltip clipping
Prior to this commit, locale was hanging out as an instance variable
in several view movels.  This meant changes to the locale from the
frontend had to be explicitly set across all those models instead
of at just one model.

With this commit, the settings model becomes /the/ holder of locale
state for the frontend.