persist app-settings to local storage
If the user changes a setting like locale from "en-US" to "es-ES", then the next time
the app fires up, the text should be rendered in Spanish, instead of defaulting back to English.

With this commit, we persist ModeSettings state to local storage (if local storage is available).
Specifically, we save current settings when the user has edited a setting and we read from
local storage when the application starts up or is reloaded.