fix regression to currency formatting when changing locale
When changing from "en-US" to "es-ES" I noticed the currency formatting
on the results page for median home price changed from:

  $ ###,###

to:

  ###.### US$

/only/ if I reloaded the app.

The Intl.Number formatter is only sensitive to locale and currency type
at initial object create.  It /should/ be sensitive to state changes to
either locale or currency type.

For now, I'm simply going to hardcode for en-US and USD until I get
full-on currency conversion implemented.  At that time, the event handler
for setting locale or country code (which drives currency type) should
trigger the updating of the number formatter.