add separate MVC modules for the floating access button (FAB)
The implementation of the FAB has always been a bit jumbled.
With this commit, we give the FAB a nice abstraction and clean-up
how we manage the state of the FAB.

In particular, it now has a spiffy finite-state-machine with
knowledge of how to navigate from one page to another.

Interestingly, the FAB 'knows' which page currently has our focus
and so is a logical object for persisting in the event the user
wants to resume a session from the last page they left off from.