add menu button to restore default settings integrate refactored/MVC-based floating action button (FAB) persist FAB state (including last focal page) set initial focal page from persisted FAB state if available don't re-fetch app from server when home icon clicked
On the face of it, this commit seems modest from a user interface
perspective:

  * The user will only see a new 'Restore from default settings'
    option off the Menu/Settings component.

  * They may also notice the app remembers which page they were on
    when they return on a subsequent session.

In reality, much heavy lifting had to happen to support this feature
cleanly.

Namely, we need to know which view components to redraw when
settings change (since language might change from Spanish to English,
for example).

Since I'm currently using flow synchronization (as opposed to observer
synchronization), we need to know what page has focus so we can
redraw /that/.

Since the FAB drives how we navigate the app, he's a logical choice
(for now) for hosting current page state (or at least providing
a getter which can infer that).  That's why you see all the change activity
around turning the FAB into a first-class MVC component.