improve affordance of non-selectable submenu items
Some read-only text exists as priority and settings values in the menu-drawer.
Since they are not selectable, the associated background color should not
change from the prevailing color on :hover events.

Also, I'm adding a little house icon next to the cost of living figure
to intimate that cost actually reflects /housing/ cost.