refactor more eventHandlers to be managed eventHandlers
I'm concerned that I am proliferating duplicate event handlers over time.
as the user clicks around (especially on mobile) and screens are dynamically
rebuilt.

It seems that modern practice is to put responsibility for clean-up of
event handlers is upon developer.

To even do that implies some bookkeeping for the handlers that are added.

I've already got a nice singleton event handler manager that already
prevents duplicate handlers from being added, so I'm progressively
sweeping through the code and replacing calls to:

  el.addEventHandler()

to:

  this.delegatedHandlers.addEventHandler()