DRY-up the code by vectoring page body renders through createPageBody()
I'm steadily moving toward a world in which responsive desktop
views are supported.  (Desktop has the geometry to support concurrent renders of multiple
components like priority sliders /and/ results view.)

But the mobile-first ethos with it's assumption of mutually-exclusive pages is in
the DNA of the app.  This commit is a gradual step toward re-writing that assumption.

Eventually,

    this.view.createPageBody(page)

will probably morph into something like:

    this.view.renderComponent(component).