generalize the FAB event handler across all use-cases
My goal is to register /one/ tracked event handler for the
floating access button and use it in all contexts/pages.

Prior to this commit, I was adding multiple FAB event handlers
that only differed in the target page that would be rendered
when the FAB was clicked.

But target page (or goto-page) is already managed state and
so we can leverage that to create a single FAB event handler
that will work in all cases.

This sets us up nicely to enforce singleton status to the FAB's
event handler in a subsequent commit... (with the virtue of not
proliferating duplicate FAB event handlers).