register only one event handler for FAB clicks
This is a bit of refactor the ensures only one
handler will respond to a floating action button
click.

The DOM is a little squirrely in that you may
/think/ you are creating a fresh FAB between
page views, but you may still have an event
handler from the previous page lurking around.

With this commit I tap into some code that ensures
only one click handler is registered for a FAB of a given
id.