fix annoying priority-page repaint/flash behavior with bloody chicken
You know how you let a leaky faucet continue to drip for several
weeks because you just haven't had a chance to call the plumber or
do the deep dive yourself and come back from Home Depot with $117 worth
of unnecessary washers, exotic plumbing tools, and freshly cached
knowledge from the plumbing dept dude who /seems/ to be a badass but maybe
that's pyschological projection on your part?

Well, in that spirit, I've noticed a super odd and ego-challenging
nuisance repaint behavior with my share-your-priorities page over
the past few months.  It seems to happen mainly on desktop.  I'll resize
the window and then vertically scroll the page or just mouse-around and the
politics card will just /flicker/, sometimes disappearing
entirely from the screen, forcing me to reload the page.
Also, selecting the menu and clicking on 'View' or any of the other
expanding submenus seems to trigger this super annoying flicker with
bothersome consistency.

I seek the advice of Lord Google.  Nuisance repaints may be
caused by many things, especially computed CSS values (like em's
and %'s) or possibly changes in classes that drive presentation.

I let my subconcious chew on all the shizzle I read and awake with
the idea of focussing on 'mdl-layout'-related classes on the
priorities page.

Afterall, I had already dropped a fix some time back for unresponsive
menu-drawers in Material Design Lite that involved programmatically
downgrading and then upgrading elements of this class.  Perhaps that
work-around is causing a secondary side-effect?

My priorities page /is/ setting the 'mdl-layout__content' class.
What would happen if i just commented that out?

The flicker seems to go away!

Normally I would dig-in and understand /why/, but the MDL framework
is not well-supported, sadly.  And I'm okay with an expedient win
in this scenario.

I'd be ready to checkin but I notice my sticky footer for the priorities
page has now gone away too. :-(

Presumably the mdl-layout__content class is horse-whispering
sweet nothings to flexbox to achieve the desired footer at the bottom
of the viewport.

So I reach for Lea Verou's book on CSS Secrets which talks
about adding your /own/ sticky footer with fu from Philip Walton.
It mainly says to create a flex body/container with 100vh and header / main / footer
elements and with the main div having 'flex: 1' to create
the sticky footer.  This seems to give me what I want, sans flicker.

See, webdev is easy. ;-)