don't clip right nav icons on small devices with longish titles
I noticed during l10n testing that the title for Spanish is about
16 characters long.  This had the effect of nudging the right-side
icons off the screen even though there was still white-space on
the header ... all in the name of centering the title.

For smaller screens, we make a compromise and sacrifice a centered
title for visibility of title and nav icons.