refactor settings view in prep for wiring up nav lang-button
I want to wire up the nav language icon with a dropdown selection
list similar to what is already in place on the full settings page.

Clearly there is opportunity to reuse some code, but the existing
settings code is composed of one big inline template-literal of html.

This commit separates that out into separate function calls, some
of which may get reused for the nav lang-button.