wire up the lang/locale icon in the header-nav
Provide the user an easy way to change languages without having to
navigate through the menu/settings/edit/lang path.

This is really a scalability play and I just like having a flag
on the nav :-)

Yes, I know, 'merica is not a language.  But I need to
limit the contents of the dropdown to a single token ... a flag
in this case since I don't have enough fu to limit the label
results to something other than the dropdown contents.

Trying to go minimal for the moment.

This required a fair amount of refactor to implement cleanly.