make results page more resilient to loss of internet
Gray-out the map view if the internet is down when the application loads
(especially when reloading from browser cache).

Use a default photo-view city image instead of attempting to fetch images over the wire
and failing when offline.

With this commit, I check the status of internet connectivity when the controller
is instantiated.  A runtime state variable, this.isOnline, is set to either:

  true
  false
  'unknown'

accordingly.

If the user advances to the results page, I /disable/ the map-view button in
the bottom appbar since it is heavily dependent upon runtime access to the
internet.

The other views can skate by with cached results okay, so they remain enabled.
I've added a tooltip to the map button that suggests the user 'reload' when
the internet is back.

I've made the tooltip l10n-friendly, depending upon the current locale.

At some point, I may want to add /periodic/ checks for inet connectivity,
but this commit at least gets us headed in that direction.

I /did/ make the check blocking since I only check one time (at app start-up)
and I was otherwise hitting a race condition such that results would attempt to be
displayed before the inet check callback was invoked.

A better strategy might be async callback that fires every 30-seconds or so.