persist data-view state on results page to local storage
Remember what view (photo, table, chart, map) the user was
using the last time the app was used.

This has the pleasant side-effect of improving table view on mobile
especially on orientation changes.  The inclination is to rotate to
landscape mode when viewing a table, so you can see all the columns.

However, before this commit, the view would revert to photo-view
when changing orientation from landscape to portrait since we
weren't persisting active data view and the view essentially gets
remade from scratch on an orientation change.