add ability to scroll-away header & subheader on results photo-view & table-view
Adding table-view recently underscored the need to manage viewport space
especially on mobile devices in landscape orientation.

One third(!) of the vertical space was taken up by header and subheader in that scenario.

With this commit, I add the ability to transition the headers out of view
when the user scrolls downward in the results section.